pyproject.toml
[project]
name = "todolist"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.129.0",
    "psycopg2-binary>=2.9.11",
    "pydantic[email]>=2.12.5",
    "python-dotenv>=1.2.1",
    "sqlmodel>=0.0.34",
]
